Fixed errors. All tests running now

+++ server/server.py
@@ -17,13 +17,15 @@ def welcome():
 def blackboard():
 	return 'blackboard'
 
-@app.route('/blackboard/create', methods=['GET'])      #create a new blackboard und save data in the mongoDB
+@app.route('/blackboard/create', methods=['POST'])      #create a new blackboard und save data in the mongoDB
 def createBlackboard():
 	#parameters
 	name = request.args.get('name')
-	validity = request.args.get('validity')
+	validityTime = request.args.get('validityTime')
+	if name is None or validityTime is None or not validityTime.isnumeric():
+		return "", 400
 
-	status_code = db.create_blackboard(name, validity, time.time())[1]
+	status_code = db.create_blackboard(name, validityTime)[1]
 
 	if status_code == 200:
 	    return 'Blackboard updated successfully', status_code
@@ -35,8 +37,10 @@ def displayBlackboard():
 	#parameters
 	name = request.args.get('name')
 	data = request.args.get('data')
+	if name is None or data is None:
+		return "", 400
 
-	status_code = db.display_blackboard(name, data, None, time.time())[1]
+	status_code = db.display_blackboard(name, data)[1]
     
 	if status_code == 200:
 	    return 'DISPLAY_BLACKBOARD', status_code
@@ -47,7 +51,10 @@ def displayBlackboard():
 def clearBlackboard():
 	#parameters
 	name = request.args.get('name')
-	status_code = db.clear_blackboard(name, time.time())[1]
+	if name is None:
+		return "", 400
+
+	status_code = db.clear_blackboard(name)[1]
 
 	if status_code == 200:
 	    return 'Blackboard cleared successfully', status_code
@@ -58,7 +65,20 @@ def clearBlackboard():
 def readBlackboard():
 	#parameters
 	name = request.args.get('name')
+	if name is None:
+		return "", 400
+
 	result, status_code = db.read_blackboard(name)
+
+	# Blackboard not foudn
+	if result is None:
+		return "", 404
+
+	# Blackboard is empty
+	if result["content"] == "":
+		return "", 444
+
+	result = { "validity": result["validity"], "content": result["content"] }
 
 	if status_code == 200:
 		return jsonify(result), status_code
@@ -69,7 +89,16 @@ def readBlackboard():
 def getBlackboardStatus():
 	#parameters
 	name = request.args.get('name')
+	if name is None:
+		return "", 400
+
 	result, status_code = db.get_blackboard_status(name)
+
+	# Blackboard not found
+	if result is None:
+		return "", 404
+
+	result = { "timestamp": result["timestamp"].isoformat(), "validity": result["validity"], "empty": result["content"] == "" }
 
 	if status_code == 200:
 		return jsonify(result),status_code
@@ -80,9 +109,12 @@ def getBlackboardStatus():
 def listBlackboard():
 	results, status_code = db.list_blackboards()
 
+	blackboardsList = []
+	for result in results:
+		blackboardsList.append( { "name": result['name'] } )
+
 	if status_code == 200:
-	
-		return jsonify(results), status_code
+		return jsonify(blackboardsList), status_code
 	else:
 		return 'An error occurred', status_code
 
@@ -90,6 +122,9 @@ def listBlackboard():
 def deleteBlackboard():
 	#parameters
 	name = request.args.get('name')
+	if name is None:
+		return "", 400
+
 	status_code = db.delete_blackboard(name)[1]
 
 	if status_code == 200:
